author controller and service modified

+++ src/main/java/com/rest_api/fs14backend/author/AuthorController.java
@@ -1,7 +1,8 @@
 package com.rest_api.fs14backend.author;
 //
 //
-import com.rest_api.fs14backend.category.Category;
+import com.rest_api.fs14backend.book.Book;
+import com.rest_api.fs14backend.book.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,31 +16,63 @@ import java.util.UUID;
 public class AuthorController {
   @Autowired
   private AuthorService authorService;
+  
+  @Autowired
+  private BookService bookService;
 
   @GetMapping("/")
-  public List<Author> getAuthors () {
-    return authorService.getAllAuthors();
+  public ResponseEntity<List<Author>> getAuthors () {
+    return ResponseEntity.ok(authorService.getAllAuthors());
   }
+
+  
 
   
   @PostMapping("/")
-  public ResponseEntity<?> createOne(@RequestBody AuthorDTO authorDTO) throws Exception {
-    Author author=new Author();
-    author.setName(authorDTO.getName());
-    return authorService.createOne(author);
+  public ResponseEntity<?> createOne(@RequestBody AuthorDTO authorDTO) throws Exception{
+    String authorName = authorDTO.getName();
     
+    // Delegate the task of adding an author to the service class
+    boolean authorExists = authorService.checkAuthorExists(authorName);
+    if (authorExists) {
+      return ResponseEntity.badRequest().body("Author " + authorName + " already exists");
+    }
+    Author savedAuthor = authorService.addAuthor(authorName);
+    return ResponseEntity.ok(savedAuthor);
   }
   
-
+ 
   
   @DeleteMapping(value = "/{id}")
   public ResponseEntity<?> deleteAuthor(@PathVariable UUID id) throws Exception {
-    return authorService.deleteAuthor(id);
+    Author authorToDelete=authorService.findById(id);
+    Book book=bookService.ifBookHasAuthor(authorToDelete);
+    if (null != book) {
+      return ResponseEntity.status(HttpStatus.CONFLICT).body("Book exists with this author");
+    }
+    Author deletedAuthor=authorService.deleteAuthor(id);
+    if(deletedAuthor!=null){
+      return ResponseEntity.ok(deletedAuthor);
+    }else{
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Author with id " + id + " not found");
+    }
+    
   }
+ 
   
   @PutMapping(value = "/{id}")
   public ResponseEntity<?> updateAuthor(@PathVariable UUID id, @RequestBody AuthorDTO authorDTO) throws Exception {
-    return authorService.updateAuthor(id, authorDTO);
+    boolean authorExists = authorService.checkAuthorExists(authorDTO.getName());
+    if (authorExists) {
+      return ResponseEntity.badRequest().body("Author " + authorDTO.getName()+ " already exists");
+    }
+    Author updatedAuthor = authorService.updateAuthor(id, authorDTO);
+    if(updatedAuthor!=null){
+      return ResponseEntity.ok(updatedAuthor);
+    }else{
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Author with id " + id + " not found");
+    }
+    
   }
   
 }
